WallE_Phase1/requirements.txt
openai
speechrecognition
RPi.GPIO
